HDF4: Update to 4.2.6
Required update as the 4.2.5 tarballs are no longer available. Some dramatic
changes. First of all, `configure` is broken but the new CMake script works
quite nicely. No more dependency on an ancient version of JPEG which cleans
things up a lot and allows the dylibs and tools to be built.

Only bad part is that the compiler wrappers `h4cc` and `h4fc` have
disappeared. Some libraries that use HDF4 depend on these as `config` scripts.