New Formula: VirtualBox
Contains patches that allow OS X Lion to be virtualized. This is allowed by the
Lion EULA as long as the Host OS is OS X and running on Apple hardware. Lion
installation is still problematic but this step can be done using commercial
virtualization software (VMWare provides free 30 day trials).

This formula builds version 4.0.12 because I was experiencing consistent
kernel panics when I upgraded to the offical 4.1.2 binary release. This formula
has only been tested on Snow Leopard with an x86_64 kernel and may not work on
other configurations---especially Lion which may require a 4.1.x build.